Remove TaskSuccess message (#56)
Removes the TaskSuccess message and instead derives the functionality from the DatasetPublished message. This makes it more likely to place the subsequent task on the worker in question. However, it imposes the condition that DatasetPublished must be always sent for the given task, even if the output is not actually stored in shm