[cascade.low.core] output schema user friendly (#106)
* [cascade.low.core] output schema user friendly

Previously, TaskDefinition.output_schema was a dict, and we required
key-sorted order to correspond to function output order in the generator
case. Thats a bit error prone / user unfriendly. We change to a list,
assuming the user provides the keys in the sorted order instead.

This makes us rely on pydantic/pickle preserving order -- but since its
a list of string tuples, it should be true.

* Add leading zeroes to Node Outputs

---------

Co-authored-by: Harrison Cook <harrison.cook@ecmwf.int>