[cascade.gateway] add troika basis (#131)
We now support troika + local & remote connections. We dont support
troika-slurm and troika-pbs yet -- needs more work.

It is realized by adding a new `--troika_config` option to gateway,
and a `TroikaSpecs` option to the job submission -- which picks
the particular entry from the troika's config.yml. Then, the gateway's
router, instead of directly spawning a python process with the cluster,
invokes the troika with script containing the python process cluster
command.