[cascade.benchmarks] add callable testing util (#143)
we want to test that various 3rd party callables can in fact be
executed under cascade. For this purpose, we expose a simplified
interface of the executor which accepts a CallableInstance, and
runs it in a semi-faithful regime, including serde+shm roundtrip

we put it to benchmarks module, as it is quite similar to
the existing benchmarks.util content. We showcase it on a demo
function therein, as well as new test_callables.py, in both cases
with just numpy sum